Proof-of-concept integration with Flutter
I've evaluated some graphics bindings for OCaml and don't really feel
drawn to any of them, so instead I'm going to try building the browser
GUI in Flutter and use a socket to communicate between the two
processes.

This is a rough proof-of-concept showing that this can be done. Most of
this new code (aside from the Flutter scaffolding) will be deleted and
replaced over time